requirements.txt
pyaudio==0.2.14
websocket-client==1.7.0
python-dotenv==1.0.0
openai==1.12.0
requests==2.31.0
numpy==1.26.3
asyncio==3.4.3
aiohttp==3.9.3
pydub==0.25.1